Move help and clear commands to new modules
