+++ commands/help.py
@@ -1,0 +1,37 @@
+from commands.common import _get_var_description
+
+
+def cmd_help(args, state):
+    """Show help information."""
+    print("\n" + "=" * 60)
+    print("LINKER CLI - COMMAND REFERENCE")
+    print("=" * 60)
+    print("State Management:")
+    print("  set <VAR> <value>     Set a variable (URL, DOMAIN, SELECTOR, etc.)")
+    print("  show [target]         Show variables, domains, or page data")
+    print()
+    print("Data Operations:")
+    print("  load <domain> <row>   Load URL from spreadsheet")
+    print("  check                 Analyze the current URL")
+    print("  bulk_check [csv]      Process multiple pages from CSV file")
+    print("  migrate               Migrate the current URL")
+    print(
+        "  report [--force] [<domain> <row1> [<row2> ... <rowN>]]      Generate an HTML report for the page"
+    )
+    print()
+    print("Link Migration:")
+    print("  lookup <url>          Look up where a link should point on the new site")
+    print("  links                 Analyze all links on current page for migration")
+    print()
+    print("Utility:")
+    print("  help [command]        Show this help or help for specific command")
+    print("  debug [on|off]        Toggle debug output")
+    print("  open <target>         Open DSM file, current URL, or report")
+    print("  clear                 Clear the screen")
+    print("  exit, quit            Exit the application")
+    print()
+    print("Variables:")
+    for var in state.variables.keys():
+        print(f"  {var:12} - {_get_var_description(var)}")
+    print("=" * 60)
+
